zamieniono typ wyniku na int

+++ main.py
@@ -3,5 +3,5 @@ import math
 numbers = list(map(int, input().split()))
 n = numbers[0]
 k = numbers[1]
-result= math.factorial(n)/(math.factorial(k)*math.factorial(n-k))
+result= int(math.factorial(n)/(math.factorial(k)*math.factorial(n-k)))
 print(result)
